1408-dev-ontochemplant-agent: Changed datatype from Float to Double for model classes, modified the agent main class to handle exception

+++ Agents/OntoChemPlantAgent/ontochemplant-agent/src/main/java/uk/ac/cam/cares/jps/agent/model/ontochemplant/PowerConsumption.java
@@ -5,6 +5,6 @@ import lombok.Setter;
 import uk.ac.cam.cares.ogm.models.FieldAnnotation;
 
 public class PowerConsumption extends OntoChemPlantModel {
-    @Getter @Setter @FieldAnnotation("http://www.ontology-of-units-of-measure.org/resource/om-2/hasNumericalValue") protected Float PowerConsumption_GigaJoulePerYear;
+    @Getter @Setter @FieldAnnotation("http://www.ontology-of-units-of-measure.org/resource/om-2/hasNumericalValue") protected Double PowerConsumption_GigaJoulePerYear;
 
 }
